+++ src/Nyris.Extensions.AspNetCore.CacheControl/UseRequestCacheControlMiddleware.cs
@@ -5,7 +5,6 @@ using System.Linq;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc.Filters;
 using Microsoft.Extensions.Primitives;
 using Nyris.Extensions.AspNetCore.CacheControl;
 
@@ -13,25 +12,31 @@ using Nyris.Extensions.AspNetCore.CacheControl;
 namespace Microsoft.AspNetCore.Mvc;
 
 /// <summary>
-/// Enables interpretation of the <c>Cache-Control</c> request header.
+///     Enables interpretation of the <c>Cache-Control</c> request header.
 /// </summary>
-public sealed class UseRequestCacheControlAttribute : ActionFilterAttribute
+public sealed class UseRequestCacheControlMiddleware
 {
-    private static readonly Regex Whitespace = new Regex(@"\s+", RegexOptions.Compiled);
+    private static readonly Regex Whitespace = new(@"\s+", RegexOptions.Compiled);
 
-    private static readonly Regex Durations = new Regex(
+    private static readonly Regex Durations = new(
         @"^\s*(?<name>max-age|max-stale|min-fresh)\s*=\s*(?<seconds>\d+)\s*$",
         RegexOptions.Compiled | RegexOptions.CultureInvariant | RegexOptions.IgnoreCase | RegexOptions.Singleline |
         RegexOptions.ExplicitCapture);
 
-    /// <inheritdoc />
-    public override Task OnActionExecutionAsync(ActionExecutingContext context, ActionExecutionDelegate next)
+    private readonly RequestDelegate _next;
+
+    public UseRequestCacheControlMiddleware(RequestDelegate next)
+    {
+        _next = next;
+    }
+
+    public Task InvokeAsync(HttpContext context)
     {
         // TODO: Investigate whether HTTP/2 trailers need to be evaluated as well.
-        var headers = context.HttpContext.Request.Headers;
+        var headers = context.Request.Headers;
 
         var control = new RequestCacheControl();
-        context.HttpContext.Features.Set((IRequestCacheControl) control);
+        context.Features.Set((ICacheControl) control);
 
         if (TryExtractDirectives("pragma", headers, out var directives))
         {
@@ -54,7 +59,7 @@ public sealed class UseRequestCacheControlAttribute : ActionFilterAttribute
             }
         }
 
-        return base.OnActionExecutionAsync(context, next);
+        return _next(context);
     }
 
     private static void TryParsePragmaDirective(string directive, RequestCacheControl control)
@@ -114,16 +119,16 @@ public sealed class UseRequestCacheControlAttribute : ActionFilterAttribute
         }
 
         directives = values.Value
-            .SelectMany(value => value?.Split(',', StringSplitOptions.RemoveEmptyEntries) ?? ArraySegment<string>.Empty)
-            .Where(value => !string.IsNullOrWhiteSpace(value))
-            .Select(directive =>
+            .SelectMany(static value => value?.Split(',', StringSplitOptions.RemoveEmptyEntries) ?? ArraySegment<string>.Empty)
+            .Where(static value => !string.IsNullOrWhiteSpace(value))
+            .Select(static directive =>
                 Whitespace.Replace(directive, string.Empty))
             .ToHashSet();
         return true;
     }
 
     // ReSharper disable once MemberCanBePrivate.Global
-    internal sealed class RequestCacheControl : IRequestCacheControl
+    internal sealed class RequestCacheControl : ICacheControl
     {
         /// <inheritdoc />
         public bool HeaderUsed { get; set; }
@@ -155,6 +160,6 @@ public sealed class UseRequestCacheControlAttribute : ActionFilterAttribute
         internal HashSet<string> Directives { get; } = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
 
         /// <inheritdoc />
-        IReadOnlyCollection<string> IRequestCacheControl.Directives => Directives;
+        IReadOnlyCollection<string> ICacheControl.Directives => Directives;
     }
 }